refactor: more path fixes
