+++ src/modalities/logging_broker/subscriber_impl/results_subscriber.py
@@ -2,13 +2,13 @@ from pathlib import Path
 from typing import Any, Dict
 
 import rich
-import wandb
 import yaml
 from rich.console import Group
 from rich.panel import Panel
 
+import wandb
 from modalities.batch import EvaluationResultBatch
-from modalities.config.config import WandbMode
+from modalities.component_instantiation.config.config import WandbMode
 from modalities.logging_broker.messages import Message
 from modalities.logging_broker.subscriber import MessageSubscriberIF
