+++ src/modalities/training/gradient_clipping/fsdp_gradient_clipper.py
@@ -1,7 +1,7 @@
 import torch
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 
-from modalities.config.lookup_enum import LookupEnum
+from modalities.component_instantiation.config.lookup_enum import LookupEnum
 from modalities.training.gradient_clipping.gradient_clipper import GradientClipperIF
