+++ src/modalities/training/gradient_clipping/fsdp_gradient_clipper_config.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from pydantic import BaseModel, Field
 
-from modalities.config.pydanctic_if_types import PydanticPytorchModuleType
+from modalities.component_instantiation.config.pydanctic_if_types import PydanticPytorchModuleType
 from modalities.training.gradient_clipping.fsdp_gradient_clipper import GradientClippingMode
